Modified to standard package layout. 1. Set package promotion to include basic promotion data struct. promotion package does not depend on any other subpackages under promotion 2. http package holds server features 3. mysql package build an abstract layer to fulfill data storage functionality
