+++ pkg/promotion/http/filter.go
@@ -1,10 +1,12 @@
-package promotion
+package http
 
 import (
 	"errors"
 	"fmt"
 	"regexp"
 	"strings"
+
+	"github.com/readr-media/readr-restful/pkg/promotion/mysql"
 )
 
 // ListParams used to bind query parameters when listing promotion
@@ -15,6 +17,10 @@ type ListParams struct {
 
 	Status string
 	Active *int64
+
+	// Embedded SQLO in the struct
+	// ListParams could be pass through interface because of decoupled Parse()
+	o *mysql.SQLO
 }
 
 // NewListParams create a new ListParams struct, and modify it with input functions,
@@ -42,19 +48,25 @@ func (p *ListParams) validate() (err error) {
 	return nil
 }
 
-func (p *ListParams) parse() *SQLO {
+// Parse will populate the SQLO in ListParams
+func (p *ListParams) Parse() {
 
 	// Set table name = "promotions"
 	// fields = promotions.*
-	s := NewSQLO(func(s *SQLO) {
-		s.table = "promotions"
-		s.fields = append(s.fields, sqlfield{table: "promotions", pattern: `%s.%s`, fields: []string{"*"}})
+	p.o = mysql.NewSQLO(func(s *mysql.SQLO) {
+		s.Table = "promotions"
+		s.Fields = append(s.Fields, mysql.Sqlfield{Table: "promotions", Pattern: `%s.%s`, Fields: []string{"*"}})
 	})
 	if p.MaxResult != 0 && p.Page > 0 {
-		s.pagination = fmt.Sprintf(" LIMIT %d OFFSET %d", p.MaxResult, (p.Page-1)*p.MaxResult)
+		p.o.Pagination = fmt.Sprintf(" LIMIT %d OFFSET %d", p.MaxResult, (p.Page-1)*p.MaxResult)
 	}
 	if p.Sort != "" {
-		s.FormatOrderBy(p.Sort)
+		p.o.FormatOrderBy(p.Sort)
 	}
-	return s
+}
+
+// Select is a wrap for SQLO's Select()
+func (p *ListParams) Select() (query string, args []interface{}, err error) {
+	query, args, err = p.o.Select()
+	return query, args, err
 }
