+++ routes/routes.go
@@ -5,7 +5,7 @@ import (
 	"github.com/readr-media/readr-restful/pkg/asset"
 	"github.com/readr-media/readr-restful/pkg/cards"
 	"github.com/readr-media/readr-restful/pkg/mail"
-	"github.com/readr-media/readr-restful/pkg/promotion"
+	promotion "github.com/readr-media/readr-restful/pkg/promotion/http"
 	"github.com/readr-media/readr-restful/poll"
 )
